Refactor QuizEditor and QuestionList, and fix broken add/remove functions
All list items use uuidv4 for their keys to prevent rendering of stale data on deletion of input fields. To fix extra renders that erase input when adding or removing questions, a mix of useRef and useState handles the quiz data. Rerendering of the quiz should only happen now when questions or answers are added/removed.